instagramLocalServer/requirements.txt
fastapi
uvicorn[standard]
requests
instagrapi
python-dotenv
python-multipart
Pillow